# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/auction": {
            "get": {
                "description": "获取分页后的所有拍卖品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction"
                ],
                "summary": "获取所有拍卖品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "拍卖品列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/utils.PaginationQ"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "非法请求",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建一个新的拍卖品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction"
                ],
                "summary": "创建拍卖品",
                "parameters": [
                    {
                        "description": "拍卖品信息",
                        "name": "auction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auction.Auction"
                        }
                    },
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "Access-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "拍卖品id",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "400": {
                        "description": "非法请求",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有权限",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            }
        },
        "/auction/{id}": {
            "get": {
                "description": "根据ID获取拍卖品详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction"
                ],
                "summary": "获取拍卖品详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "拍卖品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "拍卖品详情",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/auction.Auction"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "非法请求",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "拍卖品不存在",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新指定ID的拍卖品信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction"
                ],
                "summary": "更新拍卖品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "拍卖品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "拍卖品信息",
                        "name": "auction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auction.Update"
                        }
                    },
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "Access-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "400": {
                        "description": "非法请求",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "禁止访问",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "拍卖品不存在",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除指定ID的拍卖品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction"
                ],
                "summary": "删除拍卖品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "拍卖品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "Access-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "400": {
                        "description": "非法请求",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "禁止访问",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "拍卖品不存在",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            }
        },
        "/auction/{id}/bid": {
            "get": {
                "description": "获取指定拍卖品的竞拍列表,支持分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction"
                ],
                "summary": "获取拍卖品的所有竞拍",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "拍卖品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/utils.PaginationQ"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "拍卖品不存在",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "为指定拍卖品创建新的竞拍信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction"
                ],
                "summary": "创建竞拍",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "拍卖品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "竞拍信息",
                        "name": "bid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auction.Bid"
                        }
                    },
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "Access-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "禁止访问",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "拍卖品不存在",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "409": {
                        "description": "竞拍冲突",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "auction.Auction": {
            "type": "object",
            "required": [
                "name",
                "status"
            ],
            "properties": {
                "bid_increment": {
                    "type": "number",
                    "example": 10
                },
                "category_id": {
                    "type": "integer",
                    "example": 101
                },
                "description": {
                    "type": "string",
                    "example": "这里是拍品描述"
                },
                "end_time": {
                    "type": "integer",
                    "example": 1681819182
                },
                "fixed_price": {
                    "type": "number",
                    "example": 100
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/image.jpg"
                },
                "name": {
                    "type": "string",
                    "example": "上拍品名"
                },
                "processed": {
                    "type": "integer",
                    "example": 0
                },
                "product_id": {
                    "type": "integer",
                    "example": 101
                },
                "product_uid": {
                    "type": "integer",
                    "example": 101
                },
                "start_price": {
                    "type": "number",
                    "example": 2.1
                },
                "start_time": {
                    "type": "integer",
                    "example": 1681398030
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "upcoming",
                        "active",
                        "sold",
                        "unsold",
                        "expired",
                        "cancelled"
                    ],
                    "example": "upcoming"
                }
            }
        },
        "auction.Bid": {
            "type": "object",
            "required": [
                "bid_price"
            ],
            "properties": {
                "bid_price": {
                    "type": "number",
                    "example": 101.1
                },
                "customer_id": {
                    "type": "integer",
                    "example": 101
                }
            }
        },
        "auction.Update": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "upcoming",
                        "active",
                        "sold",
                        "unsold",
                        "expired",
                        "cancelled"
                    ]
                }
            }
        },
...
        "resp.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "default": 1000
                },
                "msg": {
                    "type": "string"
                },
                "time_now": {
                    "type": "integer"
                }
            }
        },
        "resp.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "default": 0
                },
                "data": {},
                "time_now": {
                    "type": "integer"
                }
            }
        },
...
        "utils.PaginationQ": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "save pagination list"
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "description": "Ok   bool   `json:\"ok\"`",
                    "type": "integer"
                }
            }
        }
    }
}